test(front): move getWhitelistedProviders tests to lib/api/assistant/models.test
